validaciones

+++ artistas/eliminar_artistas.php
@@ -1,5 +1,8 @@
 <?php
-	include 'conexion.php';
+//Verificar si existe una sesion:
+session_start();
+if(isset($_SESSION['valida']) && $_SESSION['valida'] == true){
+	include '../conexion.php';
 	$id= $_POST['id'];
 	$query = "delete from artistas where artista_id='$id';";
 	$resultado = pg_query($con,$query);
@@ -8,6 +11,11 @@
 	}else{
 		echo "No se elimino";
 	}	
+
+}
+else{
+	header('Location: ../index.php?error=2');
+}
 
 
 ?>
